Now the player friendship defined in database by pleayer ids, not the player's name

+++ src/main/java/me/dawey/friends/database/FriendDatabase.java
@@ -16,12 +16,16 @@ public class FriendDatabase {
          Dao<Friendship, Integer> dao = Friends.database.getFriendshipDao();
          long friendsSince = System.currentTimeMillis();
          
-         Friendship f1 = new Friendship(pname, fname, friendsSince);
-         dao.create(f1);
+         int pId = PlayerDatabase.getPlayerId(pname);
+         int fId = PlayerDatabase.getPlayerId(fname);
          
-         Friendship f2 = new Friendship(fname, pname, friendsSince);
-         dao.create(f2);
-         
+         if (pId != -1 && fId != -1) {
+             Friendship f1 = new Friendship(pId, fId, friendsSince);
+             dao.create(f1);
+             
+             Friendship f2 = new Friendship(fId, pId, friendsSince);
+             dao.create(f2);
+         }
       } catch (SQLException var8) {
          System.out.println("Error trying to insert to the database!");
          var8.printStackTrace();
@@ -31,12 +35,17 @@ public class FriendDatabase {
    public static long getFriendsSince(String pName, String fName) {
       try {
          Dao<Friendship, Integer> dao = Friends.database.getFriendshipDao();
-         QueryBuilder<Friendship, Integer> qb = dao.queryBuilder();
-         qb.where().eq("PlayerName", pName).and().eq("FriendName", fName);
-         Friendship friendship = dao.queryForFirst(qb.prepare());
+         int pId = PlayerDatabase.getPlayerId(pName);
+         int fId = PlayerDatabase.getPlayerId(fName);
          
-         if (friendship != null) {
-            return friendship.getFriendsSince();
+         if (pId != -1 && fId != -1) {
+             QueryBuilder<Friendship, Integer> qb = dao.queryBuilder();
+             qb.where().eq("player_id", pId).and().eq("friend_id", fId);
+             Friendship friendship = dao.queryForFirst(qb.prepare());
+             
+             if (friendship != null) {
+                return friendship.getFriendsSince();
+             }
          }
       } catch (SQLException var7) {
          var7.printStackTrace();
@@ -47,45 +56,59 @@ public class FriendDatabase {
    public static boolean areFriends(String pName, String fName) {
       try {
          Dao<Friendship, Integer> dao = Friends.database.getFriendshipDao();
-         QueryBuilder<Friendship, Integer> qb = dao.queryBuilder();
-         qb.where().eq("PlayerName", pName).and().eq("FriendName", fName);
-         return qb.countOf() > 0;
+         int pId = PlayerDatabase.getPlayerId(pName);
+         int fId = PlayerDatabase.getPlayerId(fName);
+         
+         if (pId != -1 && fId != -1) {
+             QueryBuilder<Friendship, Integer> qb = dao.queryBuilder();
+             qb.where().eq("player_id", pId).and().eq("friend_id", fId);
+             return qb.countOf() > 0;
+         }
       } catch (SQLException var5) {
          var5.printStackTrace();
          return false;
       }
+      return false;
    }
 
    public static List<String> getFriends(String pname) {
       List<String> friendNames = new ArrayList<>();
-
       try {
-         Dao<Friendship, Integer> dao = Friends.database.getFriendshipDao();
-         QueryBuilder<Friendship, Integer> qb = dao.queryBuilder();
-         qb.where().eq("PlayerName", pname);
-         List<Friendship> friendships = dao.query(qb.prepare());
+         int pId = PlayerDatabase.getPlayerId(pname);
+         if (pId != -1) {
+             Dao<Friendship, Integer> dao = Friends.database.getFriendshipDao();
+             QueryBuilder<Friendship, Integer> qb = dao.queryBuilder();
+             qb.where().eq("player_id", pId);
+             List<Friendship> friendships = dao.query(qb.prepare());
 
-         for (Friendship f : friendships) {
-            friendNames.add(f.getFriendName());
+             for (Friendship f : friendships) {
+                String fName = PlayerDatabase.getPlayerName(f.getFriendId());
+                if (fName != null) {
+                    friendNames.add(fName);
+                }
+             }
          }
       } catch (SQLException var5) {
          var5.printStackTrace();
       }
-
       return friendNames;
    }
 
    public static void deleteFriend(String pname, String fname) {
       try {
          Dao<Friendship, Integer> dao = Friends.database.getFriendshipDao();
+         int pId = PlayerDatabase.getPlayerId(pname);
+         int fId = PlayerDatabase.getPlayerId(fname);
          
-         DeleteBuilder<Friendship, Integer> db1 = dao.deleteBuilder();
-         db1.where().eq("PlayerName", pname).and().eq("FriendName", fname);
-         db1.delete();
+         if (pId != -1 && fId != -1) {
+             DeleteBuilder<Friendship, Integer> db1 = dao.deleteBuilder();
+             db1.where().eq("player_id", pId).and().eq("friend_id", fId);
+             db1.delete();
 
-         DeleteBuilder<Friendship, Integer> db2 = dao.deleteBuilder();
-         db2.where().eq("PlayerName", fname).and().eq("FriendName", pname);
-         db2.delete();
+             DeleteBuilder<Friendship, Integer> db2 = dao.deleteBuilder();
+             db2.where().eq("player_id", fId).and().eq("friend_id", pId);
+             db2.delete();
+         }
       } catch (SQLException var6) {
          var6.printStackTrace();
       }
@@ -93,13 +116,17 @@ public class FriendDatabase {
 
    public static int friendCount(String pName) {
       try {
-         Dao<Friendship, Integer> dao = Friends.database.getFriendshipDao();
-         QueryBuilder<Friendship, Integer> qb = dao.queryBuilder();
-         qb.where().eq("PlayerName", pName);
-         return (int) qb.countOf();
+         int pId = PlayerDatabase.getPlayerId(pName);
+         if (pId != -1) {
+             Dao<Friendship, Integer> dao = Friends.database.getFriendshipDao();
+             QueryBuilder<Friendship, Integer> qb = dao.queryBuilder();
+             qb.where().eq("player_id", pId);
+             return (int) qb.countOf();
+         }
       } catch (SQLException var2) {
          var2.printStackTrace();
          return 0;
       }
+      return 0;
    }
 }
